Create action, action creator, and reducer for fetching a single center
