+++ client/actions/centerActions.js
@@ -7,9 +7,9 @@ export function addCenterSuccess(center) {
 export function fetchCentersSuccess(centers) {
   return { type: types.FETCH_CENTERS_SUCCESS, centers };
 }
-export function fetchSingleCenterSuccess(center) {
-  return { type: types.FETCH_SINGLE_CENTER_SUCCESS, center };
-}
+// export function fetchSingleCenterSuccess(center) {
+//   return { type: types.FETCH_SINGLE_CENTER_SUCCESS, center };
+// }
 
 export function addCenter(values) {
   return function (dispatch) {
@@ -33,14 +33,14 @@ export function fetchCenters() {
       });
   };
 }
-export function fetchSingleCenter(centerId) {
-  return function (dispatch) {
-    return CenterApi.getOne(centerId)
-      .then((response) => {
-        dispatch(fetchSingleCenterSuccess(response.data.center));
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  }
-}
+// export function fetchSingleCenter(centerId) {
+//   return function (dispatch) {
+//     return CenterApi.getOne(centerId)
+//       .then((response) => {
+//         dispatch(fetchSingleCenterSuccess(response.data.center));
+//       })
+//       .catch((error) => {
+//         console.log(error);
+//       });
+//   }
+// }
