+++ client/reducers/rootReducer.js
@@ -2,10 +2,12 @@ import { combineReducers } from 'redux';
 import session from './sessionReducer';
 import register from './registerReducer';
 import centers from './centerReducer';
+import center from './singleCenterReducer';
 
 const rootReducer = combineReducers({
   session,
   register,
-  centers
+  centers,
+  center
 });
 export default rootReducer;
